feat: Rename motion component variants ...Exaggerated to ...Relaxed
