+++ change/@fluentui-react-motion-components-preview-ce97692b-d797-47e0-ad42-4d0561a9f7da.json
@@ -1,0 +1,7 @@
+{
+  "type": "minor",
+  "comment": "Rename motion component variants ...Exaggerated to ...Relaxed",
+  "packageName": "@fluentui/react-motion-components-preview",
+  "email": "olkatruk@microsoft.com",
+  "dependentChangeType": "patch"
+}
