--- packages/react-components/react-motion-components-preview/stories/src/Collapse/CollapseExaggerated.stories.md
@@ -1,1 +1,0 @@
-The exaggerated variant of `Collapse` is available as `CollapseExaggerated` component.
