+++ packages/react-components/react-motion-components-preview/stories/src/Collapse/CollapseRelaxed.stories.md
@@ -1,0 +1,1 @@
+The relaxed variant of `Collapse` is available as `CollapseRelaxed` component.
