--- packages/react-components/react-motion-components-preview/stories/src/Fade/FadeExaggerated.stories.md
@@ -1,1 +1,0 @@
-The exaggerated variant of `Fade` is available as `FadeExaggerated` component.
