+++ packages/react-components/react-motion-components-preview/stories/src/Fade/FadeRelaxed.stories.md
@@ -1,0 +1,1 @@
+The relaxed variant of `Fade` is available as `FadeRelaxed` component.
