--- packages/react-components/react-motion-components-preview/stories/src/Scale/ScaleExaggerated.stories.md
@@ -1,1 +1,0 @@
-The exaggerated variant of `Scale` is available as `ScaleExaggerated` component.
