+++ packages/react-components/react-motion-components-preview/stories/src/Scale/ScaleRelaxed.stories.md
@@ -1,0 +1,1 @@
+The relaxed variant of `Scale` is available as `ScaleRelaxed` component.
