+++ packages/react-components/react-motion-components-preview/stories/src/Scale/index.stories.ts
@@ -3,7 +3,7 @@ import ScaleDescription from './ScaleDescription.md';
 
 export { Default } from './ScaleDefault.stories';
 export { Snappy } from './ScaleSnappy.stories';
-export { Exaggerated } from './ScaleExaggerated.stories';
+export { Relaxed } from './ScaleRelaxed.stories';
 export { Customization } from './ScaleCustomization.stories';
 
 export default {
